here lol

+++ app/src/main/java/com/example/myapplication/Revue.kt
@@ -6,6 +6,6 @@ data class Revue(
     var titre:String,
     var commentaire:String,
     val utilisateur:String,
-    var note: Float,
+    var note: Int,
     val image: Int?
 )
